SOCO: Bugfix: Fix bug in USART
Disable overrun detection (As done in ECON) as if the overrun error
happens, we get stuck in a loop of interrupts.
Also, since we use modbus over USART (which has CRC), so protocol can
itself deal with overrun issue.

But one thing to note is that it's possible for the code to be slow
enough not to be able to handle the USART incoming interrupts (i.e. too
much processing happening in other interrupts).